Add interactive map for checking max wave height
Add a map file with minimal example for showing an
interactive map and getting the max wave height for
a coordinate we get when the user clicks the map.This
is done by using dash and dash-leaflet.

Add a geo utils module with tests for normalizing
longitude.

Modify the code for reading netcdf files to always require
keyword arguments to minimize risk of confusing latitude and longitude